requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
langchain
langchain-openai
langchain-core
langchain-community
langchain-chroma
tqdm
pandas
openai
email-validator
more-itertools
python-multipart